Abstracted logic of ResearchGroupForm to an own reusable component
