+++ client/src/pages/ResearchGroupAdminPage/components/CreateResearchGroupModal.tsx
@@ -3,6 +3,7 @@ import { useForm } from '@mantine/form'
 import { useState } from 'react'
 import KeycloakUserAutocomplete from '../../../components/KeycloakUserAutocomplete.tsx/KeycloakUserAutocomplete'
 import { GLOBAL_CONFIG } from '../../../config/global'
+import ResearchGroupForm from '../../../components/ResearchGroupForm/ResearchGroupForm'
 
 interface ICreateResearchGroupModalProps {
   opened: boolean
@@ -24,79 +25,12 @@ const CreateResearchGroupModal = ({
   onClose,
   onSubmit,
 }: ICreateResearchGroupModalProps) => {
-  const form = useForm<ResearchGroupFormValues>({
-    initialValues: {
-      name: '',
-      abbreviation: '',
-      campus: '',
-      description: '',
-      websiteUrl: '',
-      headUsername: '',
-    },
-    validateInputOnChange: true,
-    validate: {
-      name: (value) => (value.length < 2 ? 'Name must be at least 2 characters' : null),
-      headUsername: (value) => (!value ? 'Please select a group head' : null),
-    },
-  })
-
-  const [headDisplayLabel, setHeadDisplayLabel] = useState('')
-
   return (
     <Modal opened={opened} onClose={onClose} title='Create Research Group' centered>
-      <form onSubmit={form.onSubmit((values) => onSubmit(values))}>
-        <Stack>
-          <TextInput
-            label='Name'
-            placeholder='e.g., Intelligent Systems'
-            withAsterisk
-            {...form.getInputProps('name')}
-          />
-
-          <KeycloakUserAutocomplete
-            selectedLabel={headDisplayLabel}
-            onSelect={(username, label) => {
-              form.setFieldValue('headUsername', username)
-              setHeadDisplayLabel(label)
-            }}
-            label='Group Head'
-            placeholder='Search by name or email...'
-            withAsterisk
-          />
-
-          <TextInput
-            label='Abbreviation'
-            placeholder='e.g., IS'
-            {...form.getInputProps('abbreviation')}
-          />
-          <Select
-            label='Campus'
-            placeholder='Select a campus'
-            data={Object.values(GLOBAL_CONFIG.research_groups_location)}
-            {...form.getInputProps('campus')}
-          />
-          <TextInput
-            label='Website'
-            type='url'
-            placeholder='https://group-website.example.com'
-            {...form.getInputProps('websiteUrl')}
-          />
-          <Textarea
-            label='Description'
-            autosize
-            minRows={3}
-            maxLength={300}
-            {...form.getInputProps('description')}
-          />
-          <Text size='xs' c='dimmed'>
-            {form.values.description.length}/300 characters
-          </Text>
-
-          <Button type='submit' fullWidth mt='md' disabled={!form.isValid()}>
-            Create Research Group
-          </Button>
-        </Stack>
-      </form>
+      <ResearchGroupForm
+        onSubmit={(values) => onSubmit(values)}
+        submitLabel='Create Research Group'
+      />
     </Modal>
   )
 }
